fspapplication_backend/requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
django-tenant-users==2.1.0
django-tenants==3.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
Faker==37.1.0
ordered-set==4.1.0
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
